changed migration name, changed datatype

+++ src/main/resources/db/migration/V3__create_contacts_table.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS contacts;
 
 CREATE TABLE contacts (
-    user_id1 bigserial PRIMARY KEY,
-    user_id2 INT,
+    user_id1 BIGSERIAL PRIMARY KEY,
+    user_id2 BIGSERIAL,
       CONSTRAINT fk_user
         FOREIGN KEY (user_id2)
         REFERENCES users(id)
